refactor: clean worker/REPL separation, eliminate duplication (#540)
- Remove `workerMain()` from worker.ts as a parallel entry point
- Remove `handleUserInput()` from WorkerSession (dispatch was duplicated from main())
- Remove `_registerCommands()` from WorkerSession (command registration moves to main())
- Add `startWorkerMode()` to index.ts (exported for tests, replaces workerMain)
- Extend `main()` with optional `workerMode` parameter so both modes share the input loop
- Worker mode hooks into main(): same command dispatch, signal handling, sentinel handling
- Add `WorkerSession.runForQuery()`, `completeCurrentTask()`, `clearSession()`, `workspaceCommandDeps`
- Export `WS_TASK_ASSIGNED` and `WS_EVENT` sentinels from worker.ts
- Update CLAUDE.md to reflect the new architecture
- Update tests to use the new public API

Closes #540
