+++ CLAUDE.md
@@ -35,9 +35,9 @@ The foreman has **zero imports from agent code** — `src/agent/` is a TUI/worke
 
 All agent/worker code lives in `src/agent/`:
 
-- **`src/agent/index.ts`** — Entry point: core agent loop. Interactive REPL (default) or worker process (`--worker-mode`). Workers connect to the foreman, receive tasks, run Claude Agent SDK sessions, and report completion.
+- **`src/agent/index.ts`** — Entry point: unified agent loop for both REPL and worker modes. `main()` runs the input loop and owns command dispatch, signal handling, and workspace lifecycle. `startWorkerMode()` (exported for tests) sets up the workspace and `WorkerSession` and calls `main()` with a `workerMode` parameter. Interactive REPL (default) or worker process (`--worker-mode`).
 - **`src/agent/display.ts`** — Display/rendering engine for the worker TUI. Colors, ANSI escapes, status bar, Claude SDK message formatting.
-- **`src/agent/worker.ts`** — WebSocket client + task lifecycle. `WorkerSession`, `WorkerStatusModel`, event classification, debouncing.
+- **`src/agent/worker.ts`** — WebSocket client and task lifecycle only. `WorkerSession` owns the WebSocket connection, foreman protocol (hello/ack/task_assigned/event_notification/task_complete/goodbye), task state, and `WorkerStatusModel`. Exposes `runForQuery(prompt)`, `completeCurrentTask()`, `clearSession()`, and `workspaceCommandDeps` for use by `main()`. No command dispatch, signal handling, or input loop. `WS_TASK_ASSIGNED` and `WS_EVENT` are exported sentinels used by the main loop. Also exports `classifyEvent`, `debounceMs`, `WorkerStatusModel`.
 - **`src/agent/commands.ts`** — Command registry: `register(name, opts)`, `lookup(name)`, `listAll(workerMode)`, `execute(name, args)`, `_reset()`. Also exports `scoped(prefix)` which returns a bound register that auto-prepends the namespace (e.g. `scoped("workspace")("create", opts)` registers `"workspace:create"`). Each command is registered with its implementation handler (`CommandHandler`). No built-in registrations at module load — commands are registered at startup by their owning modules. `HandlerResult = void | "exit" | "task-complete"` propagates signals through `execute()`. `_reset()` clears the registry for test isolation.
 - **`src/agent/input.ts`** — User input handling: `ask()`, slash command parsing (uses canonical names only via `lookup()`), `dispatchInput()` (returns `{ type: "command"; name; args }` for registry commands), autocomplete.
 - **`src/agent/workspace.ts`** — Git/npm workspace management: `Workspace` class, branch checkout, npm install, safety confirmations. Also exports `registerWorkspaceCommands(deps, workerMode?)` which registers `workspace:create/reset/remove/prune` using `scoped("workspace")`. `WorkspaceCommandDeps` has four fields: `workspace: { current }` (mutable ref), `config: { workspaceDir, repoUrl, sessionId } | undefined`, `originalCwd`, and `confirm`. Handlers call `display.print` and `process.chdir` directly.
